unread [nfc]: Move state type definitions into model.
This is a step toward encapsulating these as an internal
implementation detail of the "unreads" model, and then
changing them to a data structure we can use more efficiently.

Ideally these would go inside the `unreadModel` module/file itself,
but that doesn't work because it entangles all the unread*Reducer
modules in the `types.js` import cycle-blob.  The eventual right
solution there is probably that the all-in-one `types` module
shouldn't exist.  But this works as an immediate solution.
