+++ src/reduxTypes.js
@@ -14,15 +14,12 @@ import type { Account, Outbox } from './types';
 import type { Action } from './actionTypes';
 import type {
   Topic,
-  HuddlesUnreadItem,
   Message,
   MuteTuple,
-  PmsUnreadItem,
   CrossRealmBot,
   RealmEmojiById,
   RealmFilter,
   Stream,
-  StreamUnreadItem,
   Subscription,
   User,
   UserGroup,
@@ -32,6 +29,7 @@ import type {
 import type { Narrow } from './utils/narrow';
 import type { SessionState } from './session/sessionReducer';
 import type { PmConversationsState } from './pm-conversations/pmConversationsModel';
+import type { UnreadState } from './unread/unreadModelTypes';
 
 export type * from './actionTypes';
 
@@ -300,26 +298,6 @@ export type TypingState = {
     userIds: number[],
   },
 };
-
-// These four are fragments of UnreadState; see below.
-export type UnreadStreamsState = StreamUnreadItem[];
-export type UnreadHuddlesState = HuddlesUnreadItem[];
-export type UnreadPmsState = PmsUnreadItem[];
-export type UnreadMentionsState = number[];
-
-/**
- * A summary of (almost) all unread messages, even those we don't have.
- *
- * The initial version the server gives us for this data is `unread_msgs` in
- * the `/register` initial state, and we largely follow the structure of
- * that.  See there (in `src/api/initialDataTypes.js`) for details.
- */
-export type UnreadState = {|
-  streams: UnreadStreamsState,
-  huddles: UnreadHuddlesState,
-  pms: UnreadPmsState,
-  mentions: UnreadMentionsState,
-|};
 
 export type UserGroupsState = UserGroup[];
